requirements.txt
keras="*"
numpy="*"
matplotlib="*"
progressbar2="*"